Moved testing base class into separate file
